src/requirements.txt
certifi==2022.12.7
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dataclasses==0.6
Flask==2.2.3
idna==3.4
importlib-metadata==4.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==2.0.0
Pillow==9.5.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
requests==2.28.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.15
tzdata==2023.3
Werkzeug==2.2.3
